moves the call to the API to the submit form, adds a conditional the display of the Clock, and cleans up code for coordinate grabbing from the BarCardList
